requirements.txt
tensorflow
ultralytics
opencv-python
numpy
pandas
matplotlib
scikit-learn
pytesseract
pyttsx3
seaborn
pillow
pickle